Reduce scope of pulling OpenSSL as a dependency
When dropping `ethereum` and `erc20` features, we can drop also OpenSSL
as a dependency. Currently `web3` crate cause `openssl` to be pulled
as a dependency no matter which features are enabled or disabled.

But at least if these features are not needed, we can now drop OpenSSL
when configuring `reqwest` dependency correctly.